add Blammo.Logging.Setup module, expand Simple module
